CS: prefix a "global" variable
While this `view` should only be included from within a method in the Comment Hacks plugin, there is no guarantee it will be.
So the variable declaration should be treated like a global variable, which it would be if it wouldn't be included from within a function.

So, I've rename the variable.
The variable was also being escaped for use in an HTML attribute at declaration. This was unnecessary as wherever it is used, it was also escaped and in a better way - i.e. escaping the complete string in one go instead of escaping parts of a string in individual pieces.