QA: add basic setup for integration tests
This commit adds:
* A test bootstrap file for integration tests to be run in a WordPress context.
* A basic test case for integration tests from which all test classes should extend.
* A GitHub Actions workflow to run the integration tests.
    This script will run against all supported WP versions on a mix of supported PHP versions, both with and without multisite.
* A helper script to install the WordPress Core files needed for the tests.
* The test case class to the list of known custom test classes + a minor tweak to the CS rules for the test bootstrap file.

I've elected for the setup for the initial tests to be an integration test setup as that will make adding the first set of tests more straight-forward as WordPress native functions won't need to be mocked.

To run the tests locally, copy the `phpunit.xml.dist` file to a `phpunit.xml` file and add the following, replacing the value with the path to a clone of the WordPress core repository you have locally.
```xml
	<php>
		<env name="WP_DEVELOP_DIR" value="path/to/a/local/WP/clone"/>
	</php>
```