Clean_Emails::comment_email_headers(): bug fix for handling unexpected input types
This adds a number of additional test cases for the `Clean_Emails_Test::test_comment_email_headers()` test method, which would fail without the bug fix.

Each of these tests, essentially test the same thing: a `$message_header` input is received, but is not of the type `string`.
While under "normal" circumstances, this shouldn't ever happen, another plugin hooked into the `comment_notification_headers` filter and _doing it wrong_ could lead to this situation.

The new tests document how receiving an unexpected input type will be handled and will safeguard the bug fix which I've put in place for this.